% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or95.R
\name{OR_95CI}
\alias{OR_95CI}
\title{inverse logistic link function}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{coefficient from the fitted model}

\item{se}{standard error of the coefficient}

\item{siglevel}{significance level}

\item{roundto}{number of decimals places to round to}
}
\value{
"(OR, ORlcl, ORucl)"
}
\description{
A function that returns the odds ratios and 95% confidence intervals from a vector of coeﬀicient estimates and a vector of standard error
}
\author{
Maksim Helmann
}
