% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Toy data}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{x1}{First covariate (continuous)}
\item{x2}{Second covariate (binary)}
\item{y}{Outcome variable (binary)}
}
}
\source{
https://q.utoronto.ca/courses/369245/files/32853101?module_item_id=5962847
}
\usage{
toydata
}
\description{
Data from some artificial study with binary outcome
}
\references{
Aya Mitani (aya.mitani@utoronto.ca)
}
\keyword{datasets}
